Move stratagem actions to directory
warning: Seperate fid2record to allow the addition of a future::stream
interface in addition to iterator.
